Fix scene items dock disappearing
Would disappear when items are reordered, removed, or added because the
scroll area isn't shown again. The loop also had a bug as it would not
remove the first item. layout->rowCount() is also not reliable according
to my debugging, as it would just keep increasing, and removed items
from removing the items does not reset the row count.